pyproject.toml
[project]
name = "coe-schedule-extractor"
version = "0.1.0"
description = "Extract COE bidding schedule from PDF using AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "google-genai>=0.1.0",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.0",
    "sshtunnel>=0.4.0",
    "paramiko<3.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
